mutualise les éléments communs à l'export xlsx des scores courants et avant/après audit
