+++ supabase/functions/export_audit_score/constants.ts
@@ -1,82 +1,39 @@
-import { Alignment, Border, Style } from 'https://esm.sh/exceljs@4.3.0';
-import { makeSolidFill } from '../_shared/exportUtils.ts';
+import * as Utils from '../_shared/exportUtils.ts';
 
-export const SCORE_HEADER_LABELS = [
-  'Potentiel collectivité',
-  'Points réalisés',
-  '% réalisé',
-  'Points programmés',
-  '% programmé',
-  'statut',
-];
-
+// libellés de toutes les colonnes
 export const COLUMN_LABELS = [
   '', // identifiant action
   '', // intitulé action
   'Phase',
   'Potentiel max',
-  ...SCORE_HEADER_LABELS,
-  ...SCORE_HEADER_LABELS,
+  ...Utils.SCORE_HEADER_LABELS,
+  ...Utils.SCORE_HEADER_LABELS,
   "Champs de précision de l'état d'avancement",
   'Documents liés',
 ];
 
-/** Couleurs de remplissage des cellule(s) */
-export const Fills = {
-  grey: makeSolidFill('d8d8d8'),
-  lightgrey: makeSolidFill('f2f2f2'),
-  yellow: makeSolidFill('fffed6'),
-};
-
-// couleurs de fond des lignes par axe et sous-axe
-export const BG_COLORS: Record<number, string[]> = {
-  1: ['f7caac', 'fbe4d5'],
-  2: ['9bc1e5', 'bdd7ee'],
-  3: ['70ae47', 'a9d08e'],
-  4: ['fdd966', 'fee699'],
-  5: ['8ea9db', 'b5c6e7'],
-  6: ['9f5fce', 'bc8fdd'],
-};
-
-// couleur de fond ligne sous-sous-axe
-export const BG_COLOR3 = 'bfbfbf'; // niveau 3
-export const BG_COLOR4 = 'd8d8d8'; // niveau 4 (CAE seulement)
-
-// bordures
-export const BORDER_MEDIUM = { style: 'medium' } as Partial<Border>;
-
-// alignements
-export const ALIGN_CENTER = {
-  vertical: 'middle',
-  horizontal: 'center',
-} as Partial<Alignment>;
-export const ALIGN_LEFT_WRAP = {
-  horizontal: 'left',
-  wrapText: true,
-} as Partial<Alignment>;
-
-// styles de fonte
-export const BOLD = { bold: true };
-export const ITALIC = { italic: true };
-
-// styles pour les lignes d'en-tête
-export const HEADING1 = {
-  font: BOLD,
-  alignment: ALIGN_CENTER,
-  border: {
-    top: BORDER_MEDIUM,
-    left: BORDER_MEDIUM,
-    right: BORDER_MEDIUM,
+// index (base 1) de toutes les colonnes
+export const COL_INDEX = {
+  arbo: 1,
+  intitule: 2,
+  phase: 3,
+  points_max_referentiel: 4,
+  pre_audit: {
+    points_max_personnalises: 5,
+    points_realises: 6,
+    score_realise: 7,
+    points_programmes: 8,
+    score_programme: 9,
+    statut: 10,
   },
-} as Partial<Style>;
-
-export const HEADING2 = {
-  font: BOLD,
-  alignment: { vertical: 'middle', horizontal: 'center', wrapText: true },
-} as Partial<Style>;
-
-export const HEADING_SCORES = {
-  ...HEADING2,
-  border: { top: BORDER_MEDIUM, bottom: BORDER_MEDIUM },
-  fill: Fills.grey,
+  courant: {
+    points_max_personnalises: 11,
+    points_realises: 12,
+    score_realise: 13,
+    points_programmes: 14,
+    score_programme: 15,
+    statut: 16,
+  },
+  commentaires: 17,
+  docs: 18,
 };
