+++ supabase/functions/export_audit_score/exportXLSX.ts
@@ -1,40 +1,17 @@
 /**
  * Export des scores et statut avant/après audit au format Excel
  */
-import { Workbook, Row } from 'https://esm.sh/exceljs@4.3.0';
+import { Workbook } from 'https://esm.sh/exceljs@4.3.0';
 import { TSupabaseClient } from '../_shared/getSupabaseClient.ts';
 import { Enums } from '../_shared/typeUtils.ts';
-import {
-  setCellNumFormat,
-  FORMAT_PERCENT,
-  makeSolidFill,
-  setCellsStyle,
-  makeEmptyCells,
-  capitalize,
-} from '../_shared/exportUtils.ts';
-import {
-  ALIGN_CENTER,
-  ALIGN_LEFT_WRAP,
-  BG_COLOR3,
-  BG_COLOR4,
-  BG_COLORS,
-  BOLD,
-  BORDER_MEDIUM,
-  COLUMN_LABELS,
-  Fills,
-  HEADING1,
-  HEADING2,
-  HEADING_SCORES,
-  ITALIC,
-  SCORE_HEADER_LABELS,
-} from './constants.ts';
+import * as Utils from '../_shared/exportUtils.ts';
+import * as Constants from './constants.ts';
 import { formatActionStatut } from '../_shared/formatActionStatut.ts';
 import {
-  fetchData,
   TActionReferentiel,
-  TComparaisonScoreAudit,
   TPreuve,
-} from './fetchData.ts';
+} from '../_shared/fetchActionsReferentiel.ts';
+import { fetchData, TComparaisonScoreAudit } from './fetchData.ts';
 
 export const exportXLSX = async (
   supabaseClient: TSupabaseClient,
@@ -58,17 +35,13 @@ export const exportXLSX = async (
     `Export audit ${referentiel.toUpperCase()}`
   );
 
-  // offset sur les index de cellules : la colonne "phase" n'est présente que pour CAE
-  const colOffset = referentiel === 'cae' ? 0 : -1;
-
-  // index des colonnes
-  const colIndex = getColIndex(colOffset);
-
   // ajoute les colonnes avec une largeur par défaut et quelques exceptions
-  worksheet.columns = new Array(colIndex.docs + 1).fill({ width: 12 });
-  worksheet.getColumn(colIndex.intitule).width = 50;
-  worksheet.getColumn(colIndex.commentaires).width = 50;
-  worksheet.getColumn(colIndex.docs).width = 50;
+  worksheet.columns = new Array(Constants.COL_INDEX.docs + 1).fill({
+    width: 12,
+  });
+  worksheet.getColumn(Constants.COL_INDEX.intitule).width = 50;
+  worksheet.getColumn(Constants.COL_INDEX.commentaires).width = 50;
+  worksheet.getColumn(Constants.COL_INDEX.docs).width = 50;
 
   // génère les lignes d'en-tête
   const headerRows = [
@@ -83,22 +56,13 @@ export const exportXLSX = async (
     [],
     // en-tête du tableau de données
     [
-      ...makeEmptyCells(colIndex.pre_audit - 1),
+      ...Utils.makeEmptyCells(Constants.COL_INDEX.points_max_referentiel),
       'Proposé avant audit dans la plateforme',
-      ...makeEmptyCells(SCORE_HEADER_LABELS.length - 1),
+      ...Utils.makeEmptyCells(Utils.SCORE_HEADER_LABELS.length - 1),
       'Audité dans la plateforme',
     ],
-    getColumnLabels(referentiel),
+    Constants.COLUMN_LABELS,
   ];
-
-  // index des lignes
-  const rowIndex = {
-    // index des 2 lignes d'en-tête du tableau
-    tableHeader1: headerRows.length - 1,
-    tableHeader2: headerRows.length,
-    // index de la 1ère ligne de données
-    dataStart: headerRows.length + 1,
-  };
 
   // génère les lignes de données
   const dataRows = scores.map((score) =>
@@ -111,137 +75,130 @@ export const exportXLSX = async (
     )
   );
 
-  // ajoute les lignes à la feuille de calcul
+  // ajoute les lignes d'en-tête et de données à la feuille de calcul
   worksheet.addRows([...headerRows, ...dataRows]);
+
+  // index des lignes
+  const rowIndex = {
+    // index des 2 lignes d'en-tête du tableau
+    tableHeader1: headerRows.length - 1,
+    tableHeader2: headerRows.length,
+    // index de la 1ère ligne de données
+    dataStart: headerRows.length + 1,
+  };
 
   // fusionne certaines cellules
   worksheet.mergeCells('A1:B1'); // nom de la collectivité
   // 1ère ligne d'en-tête des scores avant/après audit
   worksheet.mergeCells(
     rowIndex.tableHeader1,
-    colIndex.pre_audit,
+    Constants.COL_INDEX.pre_audit.points_max_personnalises,
     rowIndex.tableHeader1,
-    colIndex.pre_audit + SCORE_HEADER_LABELS.length - 1
+    Constants.COL_INDEX.pre_audit.statut
   );
   worksheet.mergeCells(
     rowIndex.tableHeader1,
-    colIndex.courant,
+    Constants.COL_INDEX.courant.points_max_personnalises,
     rowIndex.tableHeader1,
-    colIndex.courant + SCORE_HEADER_LABELS.length - 1
+    Constants.COL_INDEX.courant.statut
   );
 
-  // ajoute des styles
-  worksheet.getColumn(colIndex.intitule).alignment = ALIGN_LEFT_WRAP;
-  worksheet.getColumn(colIndex.commentaires).alignment = ALIGN_LEFT_WRAP;
-  worksheet.getColumn(colIndex.docs).alignment = ALIGN_LEFT_WRAP;
-  worksheet.getColumn(colIndex.points_max_referentiel).font = ITALIC;
-  if (colIndex.phase) {
-    worksheet.getColumn(colIndex.phase).alignment = ALIGN_CENTER;
+  // ajoute des styles à certaines colonnes et cellules
+  worksheet.getColumn(Constants.COL_INDEX.intitule).alignment =
+    Utils.ALIGN_LEFT_WRAP;
+  worksheet.getColumn(Constants.COL_INDEX.commentaires).alignment =
+    Utils.ALIGN_LEFT_WRAP;
+  worksheet.getColumn(Constants.COL_INDEX.docs).alignment =
+    Utils.ALIGN_LEFT_WRAP;
+  worksheet.getColumn(Constants.COL_INDEX.points_max_referentiel).font =
+    Utils.ITALIC;
+  if (Constants.COL_INDEX.phase) {
+    worksheet.getColumn(Constants.COL_INDEX.phase).alignment =
+      Utils.ALIGN_CENTER;
   }
-  worksheet.getCell('A1').fill = Fills.grey;
-  worksheet.getCell('B2').fill = Fills.yellow;
-  worksheet.getCell('B3').fill = Fills.yellow;
+  worksheet.getCell('A1').fill = Utils.FILL.grey;
+  worksheet.getCell('B2').fill = Utils.FILL.yellow;
+  worksheet.getCell('B3').fill = Utils.FILL.yellow;
   worksheet.getCell('B3').numFmt = 'dd/mm/yyyy';
-  worksheet.getCell(rowIndex.tableHeader1, colIndex.pre_audit).style = HEADING1;
-  worksheet.getCell(rowIndex.tableHeader1, colIndex.courant).style = HEADING1;
-  setCellsStyle(
+  worksheet.getCell(
+    rowIndex.tableHeader1,
+    Constants.COL_INDEX.pre_audit.points_max_personnalises
+  ).style = Utils.HEADING1;
+  worksheet.getCell(
+    rowIndex.tableHeader1,
+    Constants.COL_INDEX.courant.points_max_personnalises
+  ).style = Utils.HEADING1;
+  Utils.setCellsStyle(
     worksheet,
     rowIndex.tableHeader2,
-    colIndex.arbo,
-    colIndex.docs,
-    HEADING2
+    Constants.COL_INDEX.arbo,
+    Constants.COL_INDEX.docs,
+    Utils.HEADING2
   );
-  setCellsStyle(
+  Utils.setCellsStyle(
     worksheet,
     rowIndex.tableHeader2,
-    colIndex.pre_audit,
-    colIndex.commentaires - 1,
-    HEADING_SCORES
+    Constants.COL_INDEX.pre_audit.points_max_personnalises,
+    Constants.COL_INDEX.commentaires - 1,
+    Utils.HEADING_SCORES
   );
   worksheet.getCell(
     rowIndex.tableHeader2,
-    colIndex.points_max_referentiel
+    Constants.COL_INDEX.points_max_referentiel
   ).font = { bold: true, italic: true };
-  worksheet.getCell(rowIndex.tableHeader2, colIndex.pre_audit).border.left =
-    BORDER_MEDIUM;
-  worksheet.getCell(rowIndex.tableHeader2, 16 + colOffset).border.right =
-    BORDER_MEDIUM;
+  worksheet.getCell(
+    rowIndex.tableHeader2,
+    Constants.COL_INDEX.pre_audit.points_max_personnalises
+  ).border.left = Utils.BORDER_MEDIUM;
+  worksheet.getCell(
+    rowIndex.tableHeader2,
+    Constants.COL_INDEX.courant.statut
+  ).border.right = Utils.BORDER_MEDIUM;
 
-  // applique les styles au lignes de données
+  // applique les styles aux lignes de données
   scores.forEach(({ action_id }, index) => {
     const r = rowIndex.dataStart + index;
     const row = worksheet.getRow(r);
 
     if (action_id === referentiel) {
       // ligne "total"
-      setCellsStyle(worksheet, r, colIndex.arbo, colIndex.docs, { font: BOLD });
+      Utils.setCellsStyle(
+        worksheet,
+        r,
+        Constants.COL_INDEX.arbo,
+        Constants.COL_INDEX.docs,
+        {
+          font: Utils.BOLD,
+        }
+      );
     } else {
       // niveau de profondeur (case plier/déplier)
       const action = actionsParId[action_id];
       if (action?.depth > 1) {
-        row.outlineLevel = action?.depth;
+        row.outlineLevel = action.depth;
       }
       // couleur de fond
-      const color = getRowColor(action, referentiel);
-      if (color) row.fill = makeSolidFill(color);
+      const color = Utils.getRowColor(action, referentiel);
+      if (color) row.fill = Utils.makeSolidFill(color);
     }
 
     // formatage numérique des points/scores
-    setCellNumFormat(row.getCell(colIndex.pre_audit - 1));
-    setScoreFormats(row, colIndex.pre_audit);
-    setScoreFormats(row, colIndex.courant);
+    Utils.setCellNumFormat(
+      row.getCell(Constants.COL_INDEX.pre_audit.points_max_personnalises - 1)
+    );
+    Utils.setScoreFormats(
+      row,
+      Constants.COL_INDEX.pre_audit.points_max_personnalises
+    );
+    Utils.setScoreFormats(
+      row,
+      Constants.COL_INDEX.courant.points_max_personnalises
+    );
   });
 
   // exporte le fichier modifié
   return workbook.xlsx.writeBuffer();
 };
-
-/** applique le formatage numérique aux colonnes points/scores à partir de l'index (base-1) donné */
-const setScoreFormats = (row: Row, colIndex: number) => {
-  setCellNumFormat(row.getCell(colIndex));
-  setCellNumFormat(row.getCell(colIndex + 1));
-  setCellNumFormat(row.getCell(colIndex + 2), FORMAT_PERCENT);
-  setCellNumFormat(row.getCell(colIndex + 3));
-  setCellNumFormat(row.getCell(colIndex + 4), FORMAT_PERCENT);
-  row.getCell(colIndex + 5).style.alignment = ALIGN_CENTER;
-};
-
-// détermine la couleur de fond d'une ligne en fonction de la profondeur dans l'arbo
-const getRowColor = (
-  action?: TActionReferentiel,
-  referentiel: Enums<'referentiel'>
-) => {
-  if (action) {
-    const { depth, identifiant } = action;
-    if (depth === 3) return BG_COLOR3;
-    if (depth === 4 && referentiel === 'cae') return BG_COLOR4;
-
-    const axe = parseInt(identifiant.split('.')[0]);
-    const colors = BG_COLORS[axe];
-    if (colors && depth <= colors.length) return colors[depth - 1];
-  }
-};
-
-// génère un index des colonnes (base 1)
-const getColIndex = (colOffset: number) => {
-  const scoreColsCount = SCORE_HEADER_LABELS.length;
-  const pre_audit = 5 + colOffset;
-  const courant = pre_audit + scoreColsCount;
-  const commentaires = courant + scoreColsCount;
-  return {
-    arbo: 1,
-    intitule: 2,
-    phase: colOffset === 0 ? 3 : undefined,
-    points_max_referentiel: 4 + colOffset,
-    pre_audit,
-    courant,
-    commentaires,
-    docs: commentaires + 1,
-  };
-};
-
-const getColumnLabels = (referentiel: Enums<'referentiel'>) =>
-  referentiel === 'cae' ? COLUMN_LABELS : COLUMN_LABELS.toSpliced(2, 1);
 
 // génère une ligne de la feuille de calcul
 const getRowValues = (
@@ -256,6 +213,8 @@ const getRowValues = (
     action_id === referentiel ? 'Total' : action?.identifiant,
     // intitulé
     action_id === referentiel ? '' : action?.nom,
+    // phase
+    Utils.capitalize(action?.phase),
 
     // points max réf.
     pre_audit.points_max_referentiel,
@@ -278,19 +237,8 @@ const getRowValues = (
 
     // commentaires et documents,
     commentaire || '',
-    formatPreuves(preuves) || '',
+    Utils.formatPreuves(preuves) || '',
   ];
-
-  // insère la colonne "phase" si nécessaire
-  if (referentiel === 'cae') {
-    values.splice(2, 0, capitalize(action?.phase));
-  }
 
   return values;
 };
-
-const formatPreuves = (preuves?: TPreuve[]) =>
-  preuves
-    ?.map((p) => p?.lien?.url || p?.fichier?.filename || null)
-    .filter((s) => !!s)
-    .join('\n');
